Refactor component structure: update import paths, enhance Ourpartners layout, and consolidate Slider functionality
