+++ src/app/Components/Extra/page.jsx
@@ -2,7 +2,17 @@ import React from 'react'
 
 export default function Givenservices() {
   return (
-    <div>
+    <div className='pt-[6%] max-w-screen-xl mx-auto'>
+      <h1>Welcom here</h1>
+      <h1>Welcom here</h1>
+      <h1>Welcom here</h1>
+      <h1>Welcom here</h1>
+      <h1>Welcom here</h1>
+      <h1>Welcom here</h1>
+      <h1>Welcom here</h1>
+      <h1>Welcom here</h1>
+      <h1>Welcom here</h1>
+      <h1>Welcom here</h1>
       <h1>Welcom here</h1>
     </div>
   )
