+++ src/app/Components/Pages/Ourpartners.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Slider from '../Extra/Slider'
+import Slider from '../Slider'
 
 export default function Ourpartners() {
   return (
