+++ src/app/Data/AddingServicesData.js
@@ -1,0 +1,31 @@
+
+import ImgOne from '../Images/GroupTwo/3.png'
+import ImgTwo from '../Images/GroupTwo/4-1.png'
+import ImgThree from '../Images/GroupTwo/5-1-2048x1147.png'
+import ImgFour from '../Images/GroupTwo/6-1-2048x1147.png'
+import ImgFive from '../Images/GroupTwo/Untitled-design-1-1.png'
+import ImgSix from '../Images/GroupTwo/Untitled-design-2-1-2048x1147.png'
+import ImgSeven from '../Images/GroupTwo/Untitled-design-3-2048x1147.png'
+import ImgEight from '../Images/GroupTwo/Untitled-design-4-1.png'
+import ImgNine from '../Images/GroupTwo/Untitled-design-5-2048x1147.png'
+import ImgTen from '../Images/GroupTwo/Untitled-design-6-1.png'
+import ImgEleven from '../Images/GroupTwo/Untitled-design-7-1-2048x1147.png'
+import ImgTwelve from '../Images/GroupTwo/doctor-patient-with-face-masks-talking-about-disease-annual-checkup-visit-examination-medic-consulting-young-woman-cabinet-while-having-protection-against-coronavirus-2048x1152.jpg'
+import ImgThirty from '../Images/GroupTwo/hospital-ward-friendly-black-head-260nw-1985507345.jpg'
+import ImgFourteen from '../Images/GroupTwo/pexels-enginakyurt-1435904.jpg'
+
+
+
+const AddingServicesData = [
+    {
+        id: 1,
+        myImg : ImgTwelve,
+        title   : "Gynaecology",
+    },
+    {
+        id: 2,
+        myImg : ImgEleven,
+        title   : "Family planning",
+    },
+    
+]
